+++ website/api_credentials.py
@@ -1,0 +1,3 @@
+id = ""
+secret = ""
+
