+++ website/templates/base.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>{% block title %}ProBit API{% endblock %}</title>
+    <link rel="stylesheet" href="/static/base.css">
+  </head>
+  <body>
+    <div class="container">{% block content %}{% endblock %}</div>
+  </body>
+</html>
+
